+++ CMakeLists.txt
@@ -6,8 +6,6 @@ set(CMAKE_CXX_STANDARD 14)
 include_directories(.)
 
 add_executable(tRPnGine
-        attackables.cpp
-        attackables.h
-        helper.cpp
-        helper.h
         main.cpp)
+
+
